+++ me.id.webverify/webverifylib/src/main/java/me/id/webverifylib/listener/IDmeRegisterAffiliationListener.java
@@ -1,9 +1,0 @@
-package me.id.webverifylib.listener;
-
-/**
- * Listener to register a new affiliation
- */
-public interface IDmeRegisterAffiliationListener {
-  void onSuccess();
-  void onError(Throwable throwable);
-}